[DynamicDelivery] Fix base module build.gradle
A dynamic module depends on a base module, not the contrary.
You need to write the dependency to play-core in the base module if you use `api`, or you define the dependency in both modules with `implementation`.

I add the information to remove the previous dependency from the base module to the new dynamic module. It's more easily understandable like this I think.